+++ RevenueCatUI/CustomerCenter/ButtonStyles.swift
@@ -131,8 +131,8 @@ struct ButtonStyles_Previews: PreviewProvider {
 
             DismissCircleButton()
         }.padding()
-            .environment(\.appearance, CustomerCenterConfigData.standardAppearance)
-            .environment(\.localization, CustomerCenterConfigData.default.localization)
+            .environment(\.appearance, CustomerCenterConfigTestData.standardAppearance)
+            .environment(\.localization, CustomerCenterConfigTestData.customerCenterData.localization)
     }
 
 }
