+++ RevenueCatUI/CustomerCenter/Views/CompatibilityLabeledContent.swift
@@ -66,8 +66,8 @@ import SwiftUI
 @available(watchOS, unavailable)
 struct CompatibilityLabeledContent<Label: View, Content: View>: View {
 
-    @ViewBuilder let label: () -> Label
-    @ViewBuilder let content: () -> Content
+    let label: () -> Label
+    let content: () -> Content
 
     var body: some View {
         if #available(iOS 16.0, *) {
